Atualizacao Controller Configuracoes

+++ src/main/java/br/ufrpe/timeshare/dados/IRepositorioBens.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface IRepositorioBens extends IRepositorio<Bem> {
 
-    List<Bem> listarBensUsuarioPorNome(String nome, int idUsuario);
+    List<Bem> listarBensUsuarioPorNome(String nome, long idUsuario);
 
     List<Bem> listarBensDisponiveis();
